+++ src/webview/direct-connect-form.ts
@@ -6,6 +6,7 @@ import {
   ConnectedState,
   instanceOfApiKeyAndSecret,
   instanceOfBasicCredentials,
+  instanceOfScramCredentials,
 } from "../clients/sidecar";
 import { CustomConnectionSpec } from "../storage/resourceManager";
 import { SslConfig } from "./ssl-config-inputs";
@@ -59,6 +60,10 @@ class DirectConnectFormViewModel extends ViewModel {
     // @ts-expect-error the types don't know which credentials are present
     return this.kafkaCreds()?.api_key || "";
   });
+  kafkaScramUsername = this.derive(() => {
+    // @ts-expect-error the types don't know which credentials are present
+    return this.kafkaCreds()?.scram_username ?? null;
+  });
   kafkaSecret = this.derive(() => {
     // if credentials are there it means there is a secret. We handle the secrets in directConnect.ts
     return this.kafkaCreds() ? "fakeplaceholdersecrethere" : "";
@@ -69,6 +74,10 @@ class DirectConnectFormViewModel extends ViewModel {
   });
   kafkaSslConfig = this.derive(() => {
     return this.spec()?.kafka_cluster?.ssl || {};
+  });
+  kafkaHash = this.derive(() => {
+    // @ts-expect-error the types don't know which credentials are present
+    return this.kafkaCreds()?.hash_algorithm ?? "SCRAM_SHA_256";
   });
 
   /** Schema Registry */
@@ -146,6 +155,7 @@ class DirectConnectFormViewModel extends ViewModel {
     if (!creds || typeof creds !== "object") return "None";
     if (instanceOfBasicCredentials(creds)) return "Basic";
     if (instanceOfApiKeyAndSecret(creds)) return "API";
+    if (instanceOfScramCredentials(creds)) return "SCRAM";
     return "None";
   }
   resetTestResults() {
@@ -288,4 +298,4 @@ export type FormConnectionType =
   | "Confluent Platform"
   | "Other";
 
-type SupportedAuthTypes = "None" | "Basic" | "API";
+type SupportedAuthTypes = "None" | "Basic" | "API" | "SCRAM";